Removed all compiler warnings (left user warnings).

+++ Xocolatl/CocoaHTTPServer/Core/Mime/MultipartFormDataParser.m
@@ -7,9 +7,9 @@
 #pragma mark log level
 
 #ifdef DEBUG
-static const int httpLogLevel = HTTP_LOG_LEVEL_WARN;
+static const NSInteger httpLogLevel = HTTP_LOG_LEVEL_WARN;
 #else
-static const int httpLogLevel = HTTP_LOG_LEVEL_WARN;
+static const NSInteger httpLogLevel = HTTP_LOG_LEVEL_WARN;
 #endif
 
 
@@ -19,15 +19,15 @@ static const int httpLogLevel = HTTP_LOG_LEVEL_WARN;
 
 
 @interface MultipartFormDataParser (private)
-+ (NSData*) decodedDataFromData:(NSData*) data encoding:(int) encoding;
++ (NSData*) decodedDataFromData:(NSData*) data encoding:(NSInteger) encoding;
 
-- (int) findHeaderEnd:(NSData*) workingData fromOffset:(int) offset;
-- (int) findContentEnd:(NSData*) data fromOffset:(int) offset;
+- (NSInteger) findHeaderEnd:(NSData*) workingData fromOffset:(NSInteger) offset;
+- (NSInteger) findContentEnd:(NSData*) data fromOffset:(NSInteger) offset;
 
-- (int) numberOfBytesToLeavePendingWithData:(NSData*) data length:(int) length encoding:(int) encoding;
-- (int) offsetTillNewlineSinceOffset:(int) offset inData:(NSData*) data;
+- (NSInteger) numberOfBytesToLeavePendingWithData:(NSData*) data length:(NSInteger) length encoding:(NSInteger) encoding;
+- (NSInteger) offsetTillNewlineSinceOffset:(NSInteger) offset inData:(NSData*) data;
 
-- (int) processPreamble:(NSData*) workingData;
+- (NSInteger) processPreamble:(NSData*) workingData;
 
 @end
 
@@ -80,11 +80,11 @@ static const int httpLogLevel = HTTP_LOG_LEVEL_WARN;
 	// currently received chunk. Before returning, we always drop all data up to offset, leaving 
 	// only unhandled for the next call
 
-    int offset = 0;
+    NSInteger offset = 0;
 
 	// don't parse data unless its size is greater then boundary length, so we couldn't
 	// misfind the boundary, if it got split into different data chunks
-	int sizeToLeavePending = boundaryData.length;
+	NSInteger sizeToLeavePending = boundaryData.length;
 
 	if( !reachedEpilogue && workingData.length <= sizeToLeavePending )  {
 		// not enough data even to start parsing.
@@ -94,7 +94,7 @@ static const int httpLogLevel = HTTP_LOG_LEVEL_WARN;
 		}
 		if( checkForContentEnd ) {
 			if(	pendingData.length >= 2 ) {
-				if( *(uint16_t*)(pendingData.bytes + offset) == 0x2D2D ) {
+				if( *(NSUInteger*)(pendingData.bytes + offset) == 0x2D2D ) {
 					// we found the multipart end. all coming next is an epilogue.
 					HTTPLogVerbose(@"MultipartFormDataParser: End of multipart message");
 					waitingForCRLF = YES;
@@ -120,7 +120,7 @@ static const int httpLogLevel = HTTP_LOG_LEVEL_WARN;
 			// the flag will be raised to check if the last part was the last one.
 			if( offset < workingData.length -1 ) {
 				char* bytes = (char*) workingData.bytes;
-				if( *(uint16_t*)(bytes + offset) == 0x2D2D ) {
+				if( *(NSUInteger*)(bytes + offset) == 0x2D2D ) {
 					// we found the multipart end. all coming next is an epilogue.
 					HTTPLogVerbose(@"MultipartFormDataParser: End of multipart message");
 					checkForContentEnd = NO;
@@ -195,7 +195,7 @@ static const int httpLogLevel = HTTP_LOG_LEVEL_WARN;
 
 			// try to find CRLFCRLF bytes in the data, which indicates header end.
 			// we won't parse header parts, as they won't be too large.
-			int headerEnd = [self findHeaderEnd:workingData fromOffset:offset];
+			NSInteger headerEnd = [self findHeaderEnd:workingData fromOffset:offset];
 			if( -1 == headerEnd ) {
 				// didn't recieve the full header yet.
 				if( !pendingData.length) {
@@ -232,17 +232,17 @@ static const int httpLogLevel = HTTP_LOG_LEVEL_WARN;
 		}
 		// after we've got the header, we try to
 		// find the boundary in the data.
-		int contentEnd = [self findContentEnd:workingData fromOffset:offset];
+		NSInteger contentEnd = [self findContentEnd:workingData fromOffset:offset];
 		
 		if( contentEnd == -1 ) {
 
 			// this case, we didn't find the boundary, so the data is related to the current part.
 			// we leave the sizeToLeavePending amount of bytes to make sure we don't include 
 			// boundary part in processed data.
-			int sizeToPass = workingData.length - offset - sizeToLeavePending;
+			NSInteger sizeToPass = workingData.length - offset - sizeToLeavePending;
 
 			// if we parse BASE64 encoded data, or Quoted-Printable data, we will make sure we don't break the format
-			int leaveTrailing = [self numberOfBytesToLeavePendingWithData:data length:sizeToPass encoding:currentEncoding];
+			NSInteger leaveTrailing = [self numberOfBytesToLeavePendingWithData:data length:sizeToPass encoding:currentEncoding];
 			sizeToPass -= leaveTrailing;
 			
 			if( sizeToPass <= 0 ) {
@@ -256,7 +256,7 @@ static const int httpLogLevel = HTTP_LOG_LEVEL_WARN;
 			NSData* decodedData = [MultipartFormDataParser decodedDataFromData:[NSData dataWithBytesNoCopy:(char*)workingData.bytes + offset length:workingData.length - offset - sizeToLeavePending freeWhenDone:NO] encoding:currentEncoding];
 			
 			if( [delegate respondsToSelector:@selector(processContent:WithHeader:)] ) {
-				HTTPLogVerbose(@"MultipartFormDataParser: Processed %d bytes of body",sizeToPass);
+				HTTPLogVerbose(@"MultipartFormDataParser: Processed %ld bytes of body",sizeToPass);
 
 				[delegate processContent: decodedData WithHeader:currentHeader];
 			}
@@ -292,13 +292,13 @@ static const int httpLogLevel = HTTP_LOG_LEVEL_WARN;
 //-----------------------------------------------------------------
 #pragma mark private methods
 
-- (int) offsetTillNewlineSinceOffset:(int) offset inData:(NSData*) data {
+- (NSInteger) offsetTillNewlineSinceOffset:(NSInteger) offset inData:(NSData*) data {
 	char* bytes = (char*) data.bytes;
-	int length = data.length;
+	NSInteger length = data.length;
 	if( offset >= length - 1 ) 
 		return -1;
 
-	while ( *(uint16_t*)(bytes + offset) != 0x0A0D ) {
+	while ( *(NSUInteger*)(bytes + offset) != 0x0A0D ) {
 		// find the trailing \r\n after the boundary. The boundary line might have any number of whitespaces before CRLF, according to rfc2046
 
 		// in debug, we might also want to know, if the file is somehow misformatted.
@@ -322,17 +322,17 @@ static const int httpLogLevel = HTTP_LOG_LEVEL_WARN;
 }
 
 
-- (int) processPreamble:(NSData*) data {
-	int offset = 0;
+- (NSInteger) processPreamble:(NSData*) data {
+	NSInteger offset = 0;
 	
 	char* boundaryBytes = (char*) boundaryData.bytes + 2; // the first boundary won't have CRLF preceding.
     char* dataBytes = (char*) data.bytes;
-    int boundaryLength = boundaryData.length - 2;
-    int dataLength = data.length;
+    NSInteger boundaryLength = boundaryData.length - 2;
+    NSInteger dataLength = data.length;
     
 	// find the boundary without leading CRLF.
     while( offset < dataLength - boundaryLength +1 ) {
-        int i;
+        NSInteger i;
         for( i = 0;i < boundaryLength; i++ ) {
             if( boundaryBytes[i] != dataBytes[offset + i] )
                 break;
@@ -345,7 +345,7 @@ static const int httpLogLevel = HTTP_LOG_LEVEL_WARN;
  	
 	if( offset == dataLength ) {
 		// the end of preamble wasn't found in this chunk
-		int sizeToProcess = dataLength - boundaryLength;
+		NSInteger sizeToProcess = dataLength - boundaryLength;
 		if( sizeToProcess > 0) {
 			if( [delegate respondsToSelector:@selector(processPreambleData:)] ) {
 				NSData* preambleData = [NSData dataWithBytesNoCopy: (char*) data.bytes length: data.length - offset - boundaryLength freeWhenDone:NO];
@@ -371,10 +371,10 @@ static const int httpLogLevel = HTTP_LOG_LEVEL_WARN;
 
 
 
-- (int) findHeaderEnd:(NSData*) workingData fromOffset:(int)offset {
+- (NSInteger) findHeaderEnd:(NSData*) workingData fromOffset:(NSInteger)offset {
     char* bytes = (char*) workingData.bytes; 
-    int inputLength = workingData.length;
-    uint16_t separatorBytes = 0x0A0D;
+    NSInteger inputLength = workingData.length;
+    NSUInteger separatorBytes = 0x0A0D;
 
 	while( true ) {
 		if(inputLength < offset + 3 ) {
@@ -390,14 +390,14 @@ static const int httpLogLevel = HTTP_LOG_LEVEL_WARN;
 }
 
 
-- (int) findContentEnd:(NSData*) data fromOffset:(int) offset {
+- (NSInteger) findContentEnd:(NSData*) data fromOffset:(NSInteger) offset {
     char* boundaryBytes = (char*) boundaryData.bytes;
     char* dataBytes = (char*) data.bytes;
-    int boundaryLength = boundaryData.length;
-    int dataLength = data.length;
+    NSInteger boundaryLength = boundaryData.length;
+    NSInteger dataLength = data.length;
     
     while( offset < dataLength - boundaryLength +1 ) {
-        int i;
+        NSInteger i;
         for( i = 0;i < boundaryLength; i++ ) {
             if( boundaryBytes[i] != dataBytes[offset + i] )
                 break;
@@ -411,14 +411,14 @@ static const int httpLogLevel = HTTP_LOG_LEVEL_WARN;
 }
 
 
-- (int) numberOfBytesToLeavePendingWithData:(NSData*) data length:(int) length encoding:(int) encoding {
+- (NSInteger) numberOfBytesToLeavePendingWithData:(NSData*) data length:(NSInteger) length encoding:(NSInteger) encoding {
 	// If we have BASE64 or Quoted-Printable encoded data, we have to be sure
 	// we don't break the format.
-	int sizeToLeavePending = 0;
+	NSInteger sizeToLeavePending = 0;
 	
 	if( encoding == contentTransferEncoding_base64 ) {	
 		char* bytes = (char*) data.bytes;
-		int i;
+		NSInteger i;
 		for( i = length - 1; i > 0; i++ ) {
 			if( * (uint16_t*) (bytes + i) == 0x0A0D ) {
 				break;
@@ -450,7 +450,7 @@ static const int httpLogLevel = HTTP_LOG_LEVEL_WARN;
 #pragma mark decoding
 
 
-+ (NSData*) decodedDataFromData:(NSData*) data encoding:(int) encoding {
++ (NSData*) decodedDataFromData:(NSData*) data encoding:(NSInteger) encoding {
 	switch (encoding) {
 		case contentTransferEncoding_base64: {
 			return [data base64Decoded]; 
@@ -474,8 +474,8 @@ static const int httpLogLevel = HTTP_LOG_LEVEL_WARN;
 
 	NSMutableData* result = [[NSMutableData alloc] initWithLength:data.length];
 	const char* bytes = (const char*) data.bytes;
-	int count = 0;
-	int length = data.length;
+	NSInteger count = 0;
+	NSInteger length = data.length;
 	while( count < length ) {
 		if( bytes[count] == '=' ) {
 			[result appendBytes:bytes length:count];
@@ -496,7 +496,7 @@ static const int httpLogLevel = HTTP_LOG_LEVEL_WARN;
 			}
 			char encodedByte = 0;
 
-			for( int i = 0; i < sizeof(hex); i++ ) {
+			for( NSInteger i = 0; i < sizeof(hex); i++ ) {
 				if( hex[i] == bytes[0] ) {
 					encodedByte += i << 4;
 				}
